+++ CHANGELOG.md
@@ -6,7 +6,14 @@
     *Giovanni Troisi*  
 ### **Copyrigth**: *Tiledesk SRL*
 
+
+### 1.17.0-rc.1
+👉 **added**: check to project profile object to dynamically hide/show action
+
 ### 1.16.1 in PROD
+
+### 1.16.1-rc.1
+👉 **bug-fixed**: connector not drowed if multiple cds-action-reply-v2 is in cascade 
 
 ### 1.16.0 in PROD